+++ examples/pagerank.rs
@@ -4,9 +4,8 @@ extern crate timely;
 use std::collections::HashMap;
 use rand::{Rng, SeedableRng, StdRng};
 
-use timely::progress::nested::product::Product;
 use timely::dataflow::{InputHandle, ProbeHandle};
-use timely::dataflow::operators::{LoopVariable, ConnectLoop, Probe};
+use timely::dataflow::operators::{Feedback, ConnectLoop, Probe};
 use timely::dataflow::operators::generic::Operator;
 use timely::dataflow::channels::pact::Exchange;
 
@@ -17,13 +16,13 @@ fn main() {
         let mut input = InputHandle::new();
         let mut probe = ProbeHandle::new();
 
-        worker.dataflow(|scope| {
+        worker.dataflow::<usize,_,_>(|scope| {
 
             // create a new input, into which we can push edge changes.
             let edge_stream = input.to_stream(scope);
 
             // create a new feedback stream, which will be changes to ranks.
-            let (handle, rank_stream) = scope.loop_variable(usize::max_value(), 1);
+            let (handle, rank_stream) = scope.feedback(1);
 
             // bring edges and ranks together!
             let changes = edge_stream.binary_frontier(
@@ -170,7 +169,7 @@ fn main() {
             input.send(((rng1.gen_range(0, nodes), rng1.gen_range(0, nodes)), 1));
         }
 
-        input.advance_to(Product::new((), 1));
+        input.advance_to(1);
 
         while probe.less_than(input.time()) {
             worker.step();
@@ -179,7 +178,7 @@ fn main() {
         for i in 1 .. 1000 {
             input.send(((rng1.gen_range(0, nodes), rng1.gen_range(0, nodes)), 1));
             input.send(((rng2.gen_range(0, nodes), rng2.gen_range(0, nodes)), -1));
-            input.advance_to(Product::new((), 1 + i));
+            input.advance_to(i + 1);
             while probe.less_than(input.time()) {
                 worker.step();
             }